Replace add-header example
