--- edge-functions/modify-headers/README.md
@@ -1,59 +1,0 @@
-# Modify Headers Example
-
-Below is the code from [pages/\_middleware.ts](pages/_middleware.ts) showing how to edit headers at the edge:
-
-```ts
-import { NextResponse } from 'next/server'
-
-export function middleware() {
-  // Store the response so we can modify its headers
-  const response = NextResponse.next()
-
-  // Set custom header
-  response.headers.set(
-    "x-modified-edge",
-    "true"
-  )
-
-  // Return response
-  return response
-}
-```
-
-## Demo
-
-https://edge-functions-modify-headers.vercel.app
-
-## How to Use
-
-You can choose from one of the following two methods to use this repository:
-
-### One-Click Deploy
-
-Deploy the example using [Vercel](https://vercel.com?utm_source=github&utm_medium=readme&utm_campaign=next-example):
-
-[![Deploy with Vercel](https://vercel.com/button)](https://vercel.com/new/git/external?repository-url=https://github.com/vercel/examples/tree/main/edge-functions/modify-headers&project-name=modify-headers&repository-name=modify-headers)
-
-### Clone and Deploy
-
-Execute [`create-next-app`](https://github.com/vercel/next.js/tree/canary/packages/create-next-app) with [npm](https://docs.npmjs.com/cli/init) or [Yarn](https://yarnpkg.com/lang/en/docs/cli/create/) to bootstrap the example:
-
-```bash
-npx create-next-app --example https://github.com/vercel/examples/tree/main/edge-functions/modify-headers modify-headers
-# or
-yarn create next-app --example https://github.com/vercel/examples/tree/main/edge-functions/modify-headers modify-headers
-```
-
-Next, run Next.js in development mode:
-
-```bash
-npm install
-npm run dev
-
-# or
-
-yarn
-yarn dev
-```
-
-Deploy it to the cloud with [Vercel](https://vercel.com/new?utm_source=github&utm_medium=readme&utm_campaign=edge-middleware-eap) ([Documentation](https://nextjs.org/docs/deployment)).
