--- edge-functions/modify-headers/package.json
@@ -1,26 +1,0 @@
-{
-  "name": "modify-headers",
-  "version": "1.0.0",
-  "repository": "https://github.com/vercel/examples.git",
-  "license": "MIT",
-  "scripts": {
-    "dev": "next dev",
-    "build": "next build",
-    "start": "next start",
-    "lint": "next lint"
-  },
-  "dependencies": {
-    "@vercel/examples-ui": "^0.2.6",
-    "next": "^12.0.10",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2"
-  },
-  "devDependencies": {
-    "@types/node": "^17.0.14",
-    "@types/react": "^17.0.38",
-    "autoprefixer": "^10.4.2",
-    "postcss": "^8.4.6",
-    "tailwindcss": "^3.0.18",
-    "typescript": "^4.5.5"
-  }
-}
